# api/generated/api/openapi.yaml
openapi: 3.0.0
info:
  title: Khoai NATS Admin API
  version: <VERSION>
servers:
- url: http://localhost:8080
paths:
  /v1/login/start:
    post:
      description: Start a login process.
      operationId: v1LoginStart
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/V1LoginStartRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1LoginStartResponse'
          description: Login start
  /v1/login/exchange:
    post:
      description: "Use for exchange an email code, jwt."
      operationId: v1LoginExchange
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/V1LoginExchangeRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1AccessToken'
          description: Login exchange response
components:
  schemas:
    V1PrincipalType:
      enum:
      - EmailCode
      - Password
      type: string
    V1AccessToken:
      description: Access Token
      example:
        organizationId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        createdAt: 2000-01-23T04:56:07.000+00:00
        jwkId: jwkId
        roles:
        - Admin
        - Admin
        revokedBy: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        principalId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        revokedAt: 2000-01-23T04:56:07.000+00:00
        principalType: EmailCode
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          format: uuid
          type: string
        principalId:
          format: uuid
          type: string
        principalType:
          $ref: '#/components/schemas/V1PrincipalType'
        organizationId:
          format: uuid
          type: string
        revokedBy:
          format: uuid
          type: string
        revokedAt:
          format: date-time
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        roles:
          items:
            $ref: '#/components/schemas/V1PrincipalRoleType'
          type: array
        jwkId:
          type: string
      required:
      - createdAt
      - id
      - organizationId
      - principalId
      - principalType
      - roles
      - updatedAt
      type: object
    V1PrincipalRoleType:
      enum:
      - Admin
      - User
      - Service
      - CustomerAdmin
      type: string
    V1LoginStartRequest:
      example:
        email: email
      properties:
        email:
          type: string
      required:
      - email
      title: V1LoginStartRequest
      type: object
    V1LoginStartResponse:
      example:
        principalType: EmailCode
      properties:
        principalType:
          $ref: '#/components/schemas/V1PrincipalType'
      title: V1LoginStartResponse
      type: object
    V1LoginExchangeRequest:
      example:
        passwordOrCode: passwordOrCode
        email: email
        principalType: EmailCode
      properties:
        email:
          type: string
        principalType:
          $ref: '#/components/schemas/V1PrincipalType'
        passwordOrCode:
          type: string
      required:
      - email
      - passwordOrCode
      - principalType
      title: V1LoginExchangeRequest
      type: object
  securitySchemes:
    BearerToken:
      scheme: bearer
      type: http
    Roles:
      openIdConnectUrl: ""
      type: openIdConnect