# api/generated/server/model_v1_login_start_response.go
/*
 * Khoai NATS Admin API
 *
 *
 * API version: <VERSION>
 * Contact: phuc dep trai (phucmapcaumieu@gmail.com)
 */

package openapi

type V1LoginStartResponse struct {

	PrincipalType V1PrincipalType `json:"principalType,omitempty"`
}